SentinelOracle - Phase 3: Create BlockchainWatcher base class (#420)
### Context and Motivation

Task in Phase 3 of the Sentinel Oracle ([Feature
Spec](https://github.com/safe-research/safenet/blob/main/features/2026_05_05_transaction_checker_oracle_v1.md#phase-3--checker-node-service-pr-3-can-be-parallelized-with-phase-2)).
Complete list of tasks can be found in the
[milestone](https://github.com/safe-research/safenet/milestone/1)

This is in perparation for the sentinel service which will have to watch
events and persist the latest indexed state. For this the same logic as
in the validator service can be used.

The logic is mostly copied, with some minor adjustment for typing that
is worth paying attention to.

---

## CLA signature

With the submission of this Pull Request, I confirm that I have read and
agree to the terms of the [Contributor License
Agreement](https://docs.safefoundation.org/cla).
