+++ validator/src/validator.ts
@@ -3,8 +3,8 @@ import type { ChainFees } from "viem";
 import { createNonceManager, privateKeyToAccount } from "viem/accounts";
 import { jsonRpc } from "viem/nonce";
 import { z } from "zod";
-import type { WatcherConfig } from "./machine/transitions/watcher.js";
 import { createValidatorService } from "./service/service.js";
+import type { WatcherConfig } from "./shared/watcher.js";
 import type { ProtocolConfig } from "./types/interfaces.js";
 import { validatorConfigSchema } from "./types/schemas.js";
 import { formatError } from "./utils/errors.js";
